UI Done

+++ src/Cryptography/PasswordAttacks/AttackResult.java
@@ -4,5 +4,5 @@ public class AttackResult {
     public boolean AttackSuccessful;
     public String CrackedPassword;
     public int Attempts;
-    public float Duration;
+    public long Duration;
 }
